[qmp-system] 802.1ad VLANs for the routing protocol are created on top of the physical interface
Previously, when an interface (i.e. eth0) was configured as "lan" and "mesh",
it was added to the LAN bridge and, on top of the virtual "br-lan" interface,
the VLAN for the routing protocol was created (i.e. lan_12). This was wrong in
some cases, for example, when another interface was in the bridge but not meant
to be meshing. Now the VLANs are created on the interfaces actually configured
as "mesh".