[qmp-system] Restart the network instead of reloading it
In some corner cases, after changing the wireless settings,
the physical [radio] interfaces come up but the logical ones holding
them don't. By restarting the network, instead of reloading it, they
appear to do. It's inconvenient, because it breaks connections for a
while, but it's safer as the configuration is surely applied.

Fixes #464

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>