[qmp-system] Do not use VLANs by default on Ethernet interfaces
Since devices started to have switched Ethernet devices (namely, eth0.1
and eth0.2), using VLANs for the routing protocols (i.e., VLAN12 for
BMX6) became a challenge. The 802.1ad (QinQ) VLANs solution was cool,
but the MTU had to be changed to fit the extra 4 bytes, and it never
quite worked well. Therefore, it seems better to not use VLAN for the
routing protocol by default.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>