[qmp-system] Fix viface for 2.4 GHz 80211s_aplan radios
On first boot, routers with a single 2.4 GHz radio had it auto-
configured as 80211s_aplan, with interface wlan0 as 80211s and
wlan0ap as AP. However, the wlan0 interface was wrongly considered
as being inside the virtual interface (viface) "lan" instead of
"mesh_w0". This was a corner case in the qmp_get_virtual_iface()
function, which ultimately made rescue_ip 169.254.x.y to be
incorrectly assigned to "br-lan" instead of "mesh_w0".

Fixes #481.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>