[qmp-system] Disable rescue addresses for good
Some network devices with switched interfaces get the same MAC on all
the ports, so IPv4 rescue addresses ended up being repeated. Given that
the interfaces will get an IPv6 link-local address as long as they have
LAN or WAN or are inside a logical interface, rescue addresses can be
removed. In any case, they would get a random IPv4 address, which was
difficult to guess and rarely used.

Improves #481.

Signed-off-by: Roger Pueyo Centelles <roger.pueyo@guifi.net>